requirements.txt
flask
flask_sqlalchemy
flask_login
flask_wtf
wtforms
email-validator
openai
